requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
pandas==2.0.3
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
boto3==1.34.0
